backend/requirements.txt
Flask
psycopg2-binary
python-dotenv
argon2-cffi
flask-cors
